refactor

+++ src/main/java/com/lemon/mdcode/repository/ChannelListRepository.java
@@ -3,7 +3,7 @@ package com.lemon.mdcode.repository;
 import com.lemon.mdcode.domain.channel.ChannelList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChannelListRepository extends JpaRepository<ChannelList, Integer> {
+public interface ChannelListRepository extends JpaRepository<ChannelList, Long> {
